requirements.txt
fastapi
uvicorn
pyjwt
cryptography
pytest
pytest-cov
requests
